Added Workshop support, and also added an optional AFK Timer that'll take a player out of the RTV threshold.
To add a workshop map to the mapcycle is easy.
"umc_mapcycle"
{
    "Category"
    {
        "<mapDisplay>@ws.ugc<mapID>" {}
    }
}

Make an entry in your umc_mapcycle.txt like you would for a normal map, except follow this format:
<mapDisplay>@ws.ugc<mapID>

Example:
ch_moonside@ws.ugc2479398858

UMC will display what you put on the left side of the @ to the players, and automatically adds the map to the workshop sync.

* THIS FEATURE IS ONLY TESTED ON TF2